Moves Position and ProMoves to 'model' package
